+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Changelog
 =========
 
+## Version 0.3.0 (TBD)
+* Updated examples to address deprecations in `image` package
+* Added `_saturating` API to provide a way to handle overflow when working with integer kernels
+
 ## Version 0.2.0 (2024-07-14)
 * Bumped dependencies
 * Updated API for new version of the `image` library